+++ packages/access-definitions/tests/Wrapping/WrapperFactories/WrapperArrayFactoryTest.php
@@ -2,16 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Tests\Wrapping\EntityFetchers;
+namespace Tests\Wrapping\WrapperFactories;
 
 use EDT\Querying\ConditionFactories\PhpConditionFactory;
 use EDT\Querying\PropertyAccessors\ReflectionPropertyAccessor;
 use EDT\Querying\Utilities\ConditionEvaluator;
 use EDT\Querying\Utilities\Sorter;
 use EDT\Querying\Utilities\TableJoiner;
+use EDT\Wrapping\Contracts\AccessException;
+use EDT\Wrapping\Contracts\Types\FilterableTypeInterface;
+use EDT\Wrapping\Contracts\Types\IdentifiableTypeInterface;
 use EDT\Wrapping\Contracts\Types\ReadableTypeInterface;
 use EDT\Wrapping\Contracts\Types\TypeInterface;
-use EDT\Wrapping\Contracts\WrapperFactoryInterface;
 use EDT\Wrapping\Utilities\PropertyPathProcessorFactory;
 use EDT\Wrapping\Utilities\PropertyReader;
 use EDT\Wrapping\Utilities\SchemaPathProcessor;
@@ -19,7 +21,6 @@ use EDT\Wrapping\Utilities\TypeAccessor;
 use EDT\Wrapping\WrapperFactories\WrapperArrayFactory;
 use EDT\Querying\ObjectProviders\PrefilledObjectProvider;
 use EDT\Wrapping\TypeProviders\PrefilledTypeProvider;
-use EDT\Wrapping\Utilities\GenericEntityFetcher;
 use ReflectionClass;
 use Tests\data\Model\Person;
 use Tests\data\Types\BirthType;
@@ -27,7 +28,7 @@ use Tests\ModelBasedTest;
 use Tests\data\Types\AuthorType;
 use Tests\data\Types\BookType;
 
-class GenericEntityFetcherTest extends ModelBasedTest
+class WrapperArrayFactoryTest extends ModelBasedTest
 {
     /**
      * @var AuthorType
@@ -49,10 +50,6 @@ class GenericEntityFetcherTest extends ModelBasedTest
      * @var ReflectionPropertyAccessor
      */
     private $propertyAccessor;
-    /**
-     * @var WrapperFactoryInterface
-     */
-    private $nonWrappingWrapperFactory;
     /**
      * @var SchemaPathProcessor
      */
@@ -83,12 +80,6 @@ class GenericEntityFetcherTest extends ModelBasedTest
             new Sorter(new TableJoiner($this->propertyAccessor)),
             $this->authors
         );
-        $this->nonWrappingWrapperFactory = new class implements WrapperFactoryInterface {
-            public function createWrapper(object $entity, ReadableTypeInterface $type): object
-            {
-                return $entity;
-            }
-        };
         $this->schemaPathProcessor = new SchemaPathProcessor(new PropertyPathProcessorFactory(), $this->typeProvider);
         $this->typeAccessor = new TypeAccessor($this->typeProvider);
         $tableJoiner = new TableJoiner($this->propertyAccessor);
@@ -104,9 +95,8 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListBackingObjectsUnrestricted(): void
     {
-        $entityFetcher = $this->createGenericEntityFetcher($this->nonWrappingWrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('man', 'pseudonym');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
         self::assertCount(1, $filteredAuthors);
         $author = array_pop($filteredAuthors);
         self::assertSame($this->authors['king'], $author);
@@ -114,10 +104,9 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListWrappersDepthZero(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(0);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('man', 'pseudonym');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->createArrayWrappers($filteredAuthors, $this->authorType, 0);
 
         $expected = [
             0 => [
@@ -135,10 +124,9 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListWrappersDepthOne(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(1);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('man', 'pseudonym');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->createArrayWrappers($filteredAuthors, $this->authorType, 1);
 
         $expected = [
             0 => [
@@ -160,10 +148,9 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListWrappersDepthNegative(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(-1);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('man', 'pseudonym');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->createArrayWrappers($filteredAuthors, $this->authorType, -1);
 
         $expected = [
             0 => [
@@ -181,10 +168,9 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListWrappersDepthTwo(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(2);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('man', 'pseudonym');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->createArrayWrappers($filteredAuthors, $this->authorType, 2);
 
         $expected = [
             0 => [
@@ -213,10 +199,9 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testListWrappersWithMapping(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(0);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
         $hasString = $this->conditionFactory->propertyHasStringContainingCaseInsensitiveValue('Oranje', 'birthCountry');
-        $filteredAuthors = $entityFetcher->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->listEntities($this->authorType, [$hasString]);
+        $filteredAuthors = $this->createArrayWrappers($filteredAuthors, $this->authorType, 0);
 
         $expected = [
             0 => [
@@ -234,9 +219,7 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testGetAuthorWrapper(): void
     {
-        $wrapperFactory = $this->createWrapperArrayFactory(0);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
-        $fetchedAuthor = $entityFetcher->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien');
+        $fetchedAuthor = $this->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien');
 
         $expected = [
             'name'         => 'John Ronald Reuel Tolkien',
@@ -252,8 +235,7 @@ class GenericEntityFetcherTest extends ModelBasedTest
 
     public function testGetAuthorObject(): void
     {
-        $entityFetcher = $this->createGenericEntityFetcher($this->nonWrappingWrapperFactory);
-        $fetchedAuthor = $entityFetcher->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien');
+        $fetchedAuthor = $this->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien', false);
 
         self::assertSame($this->authors['tolkien'], $fetchedAuthor);
     }
@@ -284,9 +266,7 @@ class GenericEntityFetcherTest extends ModelBasedTest
         // When fetching the AuthorType::getAccessCondition method is automatically executed, in which
         // a path uses the BookType. This automatic check must not fail due to missing availability
         // of the BookType for the requesting user.
-        $wrapperFactory = $this->createWrapperArrayFactory(0);
-        $entityFetcher = $this->createGenericEntityFetcher($wrapperFactory);
-        $fetchedAuthor = $entityFetcher->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien');
+        $fetchedAuthor = $this->getEntityByIdentifier($this->authorType,'John Ronald Reuel Tolkien');
 
         $expected = [
             'name'         => 'John Ronald Reuel Tolkien',
@@ -302,14 +282,47 @@ class GenericEntityFetcherTest extends ModelBasedTest
         return new WrapperArrayFactory($this->propertyAccessor, $this->propertyReader, $this->typeAccessor, $depth);
     }
 
-    private function createGenericEntityFetcher(WrapperFactoryInterface $wrapperFactory): GenericEntityFetcher
+    private function createArrayWrappers(array $entities, $type, int $depth): array
     {
-        return new GenericEntityFetcher(
-            $this->authorProvider,
-            $this->conditionFactory,
-            $this->schemaPathProcessor,
-            $wrapperFactory
-        );
+        $wrapper = $this->createWrapperArrayFactory($depth);
+        return array_values(array_map(static function (object $entity) use ($wrapper, $type) {
+            return $wrapper->createWrapper($entity, $type);
+        }, $entities));
+    }
+
+    private function listEntities(FilterableTypeInterface $type, array $conditions): array
+    {
+        if (!$type->isAvailable()) {
+            throw AccessException::typeNotAvailable($type);
+        }
+
+        $this->schemaPathProcessor->mapFilterConditions($type, $conditions);
+        $conditions[] = $this->schemaPathProcessor->processAccessCondition($type);
+
+        return $this->authorProvider->getEntities($conditions, [], null);
+    }
+
+    /**
+     * @param IdentifiableTypeInterface&ReadableTypeInterface $type
+     * @param non-empty-string $identifier
+     */
+    public function getEntityByIdentifier(IdentifiableTypeInterface $type, string $identifier, bool $wrap = true)
+    {
+        $identifierPath = $type->getIdentifierPropertyPath();
+        $identifierCondition = $this->conditionFactory->propertyHasValue($identifier, ...$identifierPath);
+        $entities = $this->listEntities($type, [$identifierCondition]);
+        if ($wrap) {
+            $entities = $this->createArrayWrappers($entities, $type, 0);
+        }
+
+        switch (count($entities)) {
+            case 0:
+                throw AccessException::noEntityByIdentifier($type);
+            case 1:
+                return array_pop($entities);
+            default:
+                throw AccessException::multipleEntitiesByIdentifier($type);
+        }
     }
 }