+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 
-## 0.12.22 - 2022-10-31
+## Unreleased
 
 - refactor: remove deprecated `SchemaPathProcessor::mapSortMethods`, use `SchemaPathProcessor::processDefaultSortMethods()` and `SchemaPathProcessor::mapSorting()` instead
 - refactor: remove deprecated `SchemaPathProcessor::mapConditions`, use `SchemaPathProcessor::mapFilterConditions()` and `SchemaPathProcessor::processAccessCondition()` instead
 - refactor: remove deprecated `TypeRestrictedEntityProvider`, use the individual components manually and optimize them for your use-case
 - refactor: remove deprecated `GenericEntityFetcher`, use the individual components manually and optimize them for your use-case
+
+## 0.12.22 - 2022-10-31
+
 - refactor: require `QueryBuilderPreparer` for `DoctrineOrmEntityProvider` initialization
 - refactor: remove `QueryGenerator`, `QueryBuilderPreparer` or `DoctrineOrmEntityProvider` can be used instead
 - refactor: require `JoinFinder` for `QueryBuilderPreparer` initialization